refactor: 알림 응답 데이터 모델 변경

+++ android/app/src/main/java/com/mulkkam/data/remote/model/response/notification/NotificationsResponse.kt
@@ -9,7 +9,7 @@ data class NotificationsResponse(
     @SerialName("readNotificationResponses")
     val readNotificationResponses: List<ReadNotificationResponse>,
     @SerialName("nextCursor")
-    val nextCursor: Int?,
+    val nextCursor: Long?,
 )
 
 fun NotificationsResponse.toDomain(): List<Notification> = this.readNotificationResponses.map { it.toDomain() }
